fix: include fastp in mutliqc reports (#256)
_Describe the problem or feature in addition to a link to the issues._

multiqc is hardcoded to only run modules which are enabled, which did
not include fastp. Not sure this is the best behavior and we may want to
revisit this.

Context: undocumented anywhere, but aggregate QC on the cluster was
using some obscene amounts of RAM and time, which led me to optimize the
config to only run things enabled and add some custom parsing code into
the YAML config. That work was backported into the individual QC runs
which is what should be revisited.

optimizations for multiqc are documented here -
https://docs.seqera.io/multiqc/getting_started/config#optimising-run-time

Before submitting this PR, please make sure:

- [x] You have added a few sentences describing the PR here.
- [x] You have added yourself or the appropriate individual as the
assignee.
- [x] You have added at least one relevant code reviewer to the PR.
- [x] The code passes all CI tests without any errors or warnings.
- [ ] You have added tests (when appropriate).
- [ ] You have added an entry in any relevant CHANGELOGs (when
appropriate).
- [ ] If you have made any changes to the `scripts/` or `docker/`
directories, please ensure any image versions have been incremented
accordingly!
- [ ] You have updated the README or other documentation to account for
these changes (when appropriate).